Python311

+++ base/daemon_thread.py
@@ -1,4 +1,3 @@
-import json
 import logging
 import re
 import signal
@@ -8,20 +7,19 @@ import time
 from dataclasses import dataclass
 from datetime import datetime, timedelta
 from pprint import pformat
-from typing import Dict, List, Union
 
 from kubernetes import client, watch
 from kubernetes import config as kube_config
-from kubernetes.client import ApiClient
-from pyzabbix import ZabbixMetric, ZabbixSender
+from kubernetes.client import ApiClient, CoreV1Api, AppsV1Api, ExtensionsV1beta1Api
+from pyzabbix import ZabbixMetric, ZabbixSender, ZabbixResponse
 
 from base.config import Configuration, ClusterAccessConfigType
 from base.timed_threads import TimedThread
 from base.watcher_thread import WatcherThread
-from k8sobjects import K8sObject
-from k8sobjects.container import get_container_zabbix_metrics
+from k8sobjects.k8sobject import K8sObject
 from k8sobjects.k8sresourcemanager import K8sResourceManager
 from k8sobjects.pvc import get_pvc_volumes_for_all_nodes
+from .web_api import WebApi
 
 exit_flag = threading.Event()
 
@@ -56,15 +54,15 @@ class KubernetesApi:
 
 
 class CheckKubernetesDaemon:
-    data: Dict[str, K8sResourceManager] = {}
-    discovery_sent: Dict[str, datetime] = {}
+    data: dict[str, K8sResourceManager] = {}
+    discovery_sent: dict[str, datetime] = {}
     thread_lock = threading.Lock()
 
     def __init__(self, config: Configuration,
-                 resources: List[str],
+                 resources: list[str],
                  discovery_interval: int, data_resend_interval: int,
                  ):
-        self.manage_threads: List[Union[TimedThread, WatcherThread]] = []
+        self.manage_threads: list[TimedThread | WatcherThread] = []
         self.config = config
         self.logger = logging.getLogger(__file__)
         self.discovery_interval = int(discovery_interval)
@@ -122,14 +120,14 @@ class CheckKubernetesDaemon:
             self.logger.info(f"WEB Api Host {self.web_api_host} with resources {','.join(self.web_api_resources)}")
 
     @staticmethod
-    def exclude_resources(available_types: List[str], excluded_types: List[str]) -> List[str]:
+    def exclude_resources(available_types: list[str], excluded_types: list[str]) -> list[str]:
         result = []
         for k8s_type_available in available_types:
             if k8s_type_available not in excluded_types:
                 result.append(k8s_type_available)
         return result
 
-    def handler(self, signum: int, *args: str):
+    def handler(self, signum: int, *args: str) -> None:
         if signum in [signal.SIGTERM]:
             self.logger.info('Signal handler called with signal %s... stopping (max %s seconds)' % (signum, 3))
             exit_flag.set()
@@ -157,13 +155,14 @@ class CheckKubernetesDaemon:
                         data_print = pformat(obj_d.data, indent=2)
                         self.logger.info(f"resource={r}, object_name={obj_name}, object_data={data_print}")
 
-    def run(self):
+    def run(self) -> None:
         self.start_data_threads()
         self.start_api_info_threads()
         self.start_loop_send_discovery_threads()
         self.start_resend_threads()
 
-    def start_data_threads(self):
+    def start_data_threads(self) -> None:
+        thread: WatcherThread | TimedThread
         for resource in self.resources:
             with self.thread_lock:
                 self.data.setdefault(resource, K8sResourceManager(resource, zabbix_host=self.zabbix_host))
@@ -183,14 +182,8 @@ class CheckKubernetesDaemon:
                                      daemon_object=self, daemon_method='watch_data')
                 self.manage_threads.append(thread)
                 thread.start()
-            else:
-                thread = WatcherThread(resource, exit_flag,
-                                       daemon_object=self, daemon_method='watch_data')
-                self.manage_threads.append(thread)
-                thread.start()
-
             # additional looping data threads
-            if resource == 'services':
+            elif resource == 'services':
                 thread = TimedThread(resource, self.data_resend_interval, exit_flag,
                                      daemon_object=self, daemon_method='report_global_data_zabbix',
                                      delay_first_run_seconds=self.discovery_interval + 5)
@@ -202,8 +195,13 @@ class CheckKubernetesDaemon:
                                      delay_first_run_seconds=self.discovery_interval + 5)
                 self.manage_threads.append(thread)
                 thread.start()
+            else:
+                thread = WatcherThread(resource, exit_flag,
+                                       daemon_object=self, daemon_method='watch_data')
+                self.manage_threads.append(thread)
+                thread.start()
 
-    def start_api_info_threads(self):
+    def start_api_info_threads(self) -> None:
         if 'nodes' not in self.resources:
             # only send api heartbeat once
             return
@@ -213,7 +211,7 @@ class CheckKubernetesDaemon:
         self.manage_threads.append(thread)
         thread.start()
 
-    def start_loop_send_discovery_threads(self):
+    def start_loop_send_discovery_threads(self) -> None:
         for resource in self.resources:
             send_discovery_thread = TimedThread(resource, self.discovery_interval, exit_flag,
                                                 daemon_object=self, daemon_method='send_zabbix_discovery',
@@ -222,7 +220,7 @@ class CheckKubernetesDaemon:
             self.manage_threads.append(send_discovery_thread)
             send_discovery_thread.start()
 
-    def start_resend_threads(self):
+    def start_resend_threads(self) -> None:
         for resource in self.resources:
             resend_thread = TimedThread(resource, self.data_resend_interval, exit_flag,
                                         daemon_object=self, daemon_method='resend_data',
@@ -232,7 +230,7 @@ class CheckKubernetesDaemon:
             self.manage_threads.append(resend_thread)
             resend_thread.start()
 
-    def get_api_for_resource(self, resource: str):
+    def get_api_for_resource(self, resource: str) -> CoreV1Api | AppsV1Api | ExtensionsV1beta1Api:
         if resource in ['nodes', 'components', 'secrets', 'pods', 'services', 'pvcs']:
             api = self.core_v1
         elif resource in ['deployments', 'daemonsets', 'statefulsets']:
@@ -243,13 +241,12 @@ class CheckKubernetesDaemon:
             raise AttributeError('No valid resource found: %s' % resource)
         return api
 
-    def get_web_api(self):
+    def get_web_api(self) -> WebApi:
         if not hasattr(self, '_web_api'):
-            from .web_api import WebApi
             self._web_api = WebApi(self.web_api_host, self.web_api_token, verify_ssl=self.web_api_verify_ssl)
         return self._web_api
 
-    def watch_data(self, resource: str):
+    def watch_data(self, resource: str) -> None:
         api = self.get_api_for_resource(resource)
         stream_named_arguments = {"timeout_seconds": self.config.k8s_api_stream_timeout_seconds}
         request_named_arguments = {"_request_timeout": self.config.k8s_api_request_timeout_seconds}
@@ -305,7 +302,7 @@ class CheckKubernetesDaemon:
                 time.sleep(60)
             self.logger.debug("Watch/fetch completed for resource >>>%s<<<, restarting" % resource)
 
-    def watch_event_handler(self, resource: str, event: Dict):
+    def watch_event_handler(self, resource: str, event: dict) -> None:
 
         obj = event['object'].to_dict()
         event_type = event['type']
@@ -344,14 +341,15 @@ class CheckKubernetesDaemon:
         elif event_type.lower() == 'deleted':
             with self.thread_lock:
                 resourced_obj = self.data[resource].del_obj(obj)
-                self.delete_object(resource, resourced_obj)
+                if resourced_obj:
+                    self.delete_object(resource, resourced_obj)
         else:
             self.logger.info('event type "%s" not implemented' % event_type)
 
-    def report_global_data_zabbix(self, resource):
+    def report_global_data_zabbix(self, resource: str) -> None:
         """ aggregate and report information for some speciality in resources """
         if resource not in self.discovery_sent:
-            self.logger.debug('skipping report_global_data_zabbix for %s, disovery not send yet!' % resource)
+            self.logger.debug('skipping report_global_data_zabbix for %s, discovery not send yet!' % resource)
             return
 
         data_to_send = list()
@@ -372,43 +370,48 @@ class CheckKubernetesDaemon:
                 ZabbixMetric(self.zabbix_host, 'check_kubernetes[get,services,num_ingress_services]',
                              str(num_ingress_services)))
             self.send_data_to_zabbix(resource, None, data_to_send)
-        elif resource == 'containers':
-            # aggregate pod data to containers for each namespace
-            with self.thread_lock:
-                containers = dict()
-                for obj_uid, resourced_obj in self.data['pods'].objects.items():
-                    ns = resourced_obj.name_space
-                    if ns not in containers:
-                        containers[ns] = dict()
 
-                    pod_data = resourced_obj.resource_data
-                    pod_base_name = resourced_obj.base_name
-                    try:
-                        container_status = json.loads(pod_data['container_status'])
-                    except Exception as e:
-                        self.logger.error(e)
-                        continue
+        # TODO: disabled, rewrite later
+        # elif resource == 'containers':
+        #     # aggregate pod data to containers for each namespace
+        #     with self.thread_lock:
+        #         containers: dict[str, dict[str, dict[str, object]]] = dict()
+        #         for obj_uid, resourced_obj in self.data['pods'].objects.items():
+        #             if resourced_obj.name_space is None:
+        #                 continue
+        #             if not isinstance(resourced_obj, Pod):
+        #                 continue
+        #             ns = resourced_obj.name_space
+        #             containers.setdefault(resourced_obj.name_space, dict())
+        #
+        #             pod_data = resourced_obj.resource_data
+        #             pod_base_name = resourced_obj.base_name
+        #             try:
+        #                 container_status: dict[str, dict[str, object]] = json.loads(pod_data['container_status'])
+        #             except Exception as e:
+        #                 self.logger.error(e)
+        #                 continue
+        #
+        #             # aggregate container information
+        #             for container_name, container_data in container_status.items():
+        #                 containers[ns].setdefault(pod_base_name, dict())
+        #                 containers[ns][pod_base_name].setdefault(container_name, container_data)
+        #
+        #                 for k, v in containers[ns][pod_base_name][container_name].items():
+        #                     if isinstance(v, int):
+        #                         containers[ns][pod_base_name][container_name][k] += container_data[k]
+        #                     elif k == 'status' and container_data[k].startswith('ERROR'):
+        #                         containers[ns][pod_base_name][container_name][k] = container_data[k]
+        #
+        #         for ns, d1 in containers.items():
+        #             for pod_base_name, d2 in d1.items():
+        #                 for container_name, container_data in d2.items():
+        #                     data_to_send += get_container_zabbix_metrics(self.zabbix_host, ns, pod_base_name,
+        #                                                                  container_name, container_data)
+        #
+        #         self.send_data_to_zabbix(resource, None, data_to_send)
 
-                    # aggregate container information
-                    for container_name, container_data in container_status.items():
-                        containers[ns].setdefault(pod_base_name, dict())
-                        containers[ns][pod_base_name].setdefault(container_name, container_data)
-
-                        for k, v in containers[ns][pod_base_name][container_name].items():
-                            if isinstance(v, int):
-                                containers[ns][pod_base_name][container_name][k] += container_data[k]
-                            elif k == 'status' and container_data[k].startswith('ERROR'):
-                                containers[ns][pod_base_name][container_name][k] = container_data[k]
-
-                for ns, d1 in containers.items():
-                    for pod_base_name, d2 in d1.items():
-                        for container_name, container_data in d2.items():
-                            data_to_send += get_container_zabbix_metrics(self.zabbix_host, ns, pod_base_name,
-                                                                         container_name, container_data)
-
-                self.send_data_to_zabbix(resource, None, data_to_send)
-
-    def resend_data(self, resource):
+    def resend_data(self, resource: str) -> None:
 
         with self.thread_lock:
             try:
@@ -456,11 +459,11 @@ class CheckKubernetesDaemon:
             except RuntimeError as e:
                 self.logger.warning(str(e))
 
-    def delete_object(self, resource_type, resourced_obj):
+    def delete_object(self, resource_type: str, resourced_obj: K8sObject) -> None:
         # TODO: trigger zabbix discovery, srsly?
         self.send_to_web_api(resource_type, resourced_obj, "deleted")
 
-    def send_zabbix_discovery(self, resource: str):
+    def send_zabbix_discovery(self, resource: str) -> None:
         # aggregate data and send to zabbix
         self.logger.info(f"send_zabbix_discovery: {resource}")
         with self.thread_lock:
@@ -481,7 +484,9 @@ class CheckKubernetesDaemon:
 
             self.discovery_sent[resource] = datetime.now()
 
-    def send_object(self, resource, resourced_obj, event_type, send_zabbix_data=False, send_web=False):
+    def send_object(self, resource: str, resourced_obj: K8sObject,
+                    event_type: str, send_zabbix_data: bool = False,
+                    send_web: bool = False) -> None:
         # send single object for updates
         with self.thread_lock:
             if send_zabbix_data:
@@ -506,16 +511,16 @@ class CheckKubernetesDaemon:
                         resource, resourced_obj.name_space, resourced_obj.name, self.rate_limit_seconds))
                     resourced_obj.is_dirty_web = True
 
-    def send_heartbeat_info(self, *args):
+    def send_heartbeat_info(self, resource: str) -> None:
         result = self.send_to_zabbix([
             ZabbixMetric(self.zabbix_host, 'check_kubernetesd[discover,api]', str(int(time.time())))
         ])
         if result.failed > 0:
-            self.logger.error("failed to send heartbeat to zabbix")
+            self.logger.error(f"{resource} failed to send heartbeat to zabbix")
         else:
-            self.logger.debug("successfully sent heartbeat to zabbix ")
+            self.logger.debug(f"{resource} successfully sent heartbeat to zabbix ")
 
-    def send_to_zabbix(self, metrics: List[ZabbixMetric]):
+    def send_to_zabbix(self, metrics: list[ZabbixMetric]) -> ZabbixResponse | DryResult:
         if self.zabbix_dry_run:
             result = DryResult()
         else:
@@ -534,20 +539,20 @@ class CheckKubernetesDaemon:
                 self.logger.info('===> Sending to zabbix: >>>%s<<<' % metrics)
         return result
 
-    def send_discovery_to_zabbix(self, resource: str, metric: ZabbixMetric = None, obj: K8sObject = None):
+    def send_discovery_to_zabbix(self, resource: str, metric: ZabbixMetric = None, obj: K8sObject = None) -> None:
         if resource not in self.zabbix_resources:
             self.logger.warning(
                 f'resource {resource} ist not activated, active resources are : {",".join(self.zabbix_resources)}')
             return
 
         if obj:
-            discovery_data = obj.get_discovery_for_zabbix()
+            discovery_data = obj.get_discovery_for_zabbix(metric)
             if not discovery_data:
                 self.logger.warning('No discovery_data for obj %s, not sending to zabbix!' % obj.uid)
                 return
 
             discovery_key = 'check_kubernetesd[discover,' + resource + ']'
-            result = self.send_to_zabbix([ZabbixMetric(self.zabbix_host, discovery_key, discovery_data)])
+            result = self.send_to_zabbix([ZabbixMetric(host=self.zabbix_host, key=discovery_key, value=discovery_data)])
             if result.failed > 0:
                 self.logger.error("failed to sent zabbix discovery: %s : >>>%s<<<" % (discovery_key, discovery_data))
             elif self.zabbix_debug:
@@ -561,7 +566,8 @@ class CheckKubernetesDaemon:
         else:
             self.logger.warning('No obj or metrics found for send_discovery_to_zabbix [%s]' % resource)
 
-    def send_data_to_zabbix(self, resource: str, obj=None, metrics=None):
+    def send_data_to_zabbix(self, resource: str, obj: K8sObject = None,
+                            metrics: list[ZabbixMetric] | None = None) -> None:
         if metrics is None:
             metrics = list()
         if resource not in self.zabbix_resources:
@@ -594,7 +600,7 @@ class CheckKubernetesDaemon:
                 self.logger.debug("successfully sent %s zabbix items [%s: %s]" % (
                     len(metrics), resource, obj.name if obj else 'metrics'))
 
-    def send_to_web_api(self, resource, obj, action):
+    def send_to_web_api(self, resource: str, obj: K8sObject, action: str) -> None:
         if resource not in self.web_api_resources:
             return
